pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-syauth"
version = "0.1.0"
description = "Official Django SDK for Syauth Authentication"
readme = "README.md"
authors = [
  { name = "Nexorix", email = "support@nexorix.com" },
]
license = { file = "LICENSE" }
keywords = ["django", "authentication", "oauth", "oidc", "syauth", "sso"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Session",
    "Topic :: Security",
]
dependencies = [
    "django>=4.0",
    "requests>=2.25.0",
    "pyjwt[crypto]>=2.8.0",
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://syauth.com"
Documentation = "https://developers.nexorix.com/syauth"
Repository = "https://github.com/nexorix/django-syauth"
Issues = "https://github.com/nexorix/django-syauth/issues"
Changelog = "https://github.com/nexorix/django-syauth/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]
